move controlsRow generation to ChartTab
