+++ grapher/chart/ChartTab.tsx
@@ -1,5 +1,5 @@
 import * as React from "react"
-import { computed } from "mobx"
+import { action, computed } from "mobx"
 import { observer } from "mobx-react"
 import { SlopeChart } from "grapher/slopeCharts/SlopeChart"
 import { Bounds } from "grapher/utils/Bounds"
@@ -13,6 +13,17 @@ import { StackedBarChart } from "grapher/barCharts/StackedBarChart"
 import { ChartLayout, ChartLayoutView } from "./ChartLayout"
 import { TimeScatter } from "grapher/scatterCharts/TimeScatter"
 import { LoadingOverlay } from "grapher/loadingIndicator/LoadingOverlay"
+import { faExchangeAlt } from "@fortawesome/free-solid-svg-icons/faExchangeAlt"
+import { faPencilAlt } from "@fortawesome/free-solid-svg-icons/faPencilAlt"
+import { faPlus } from "@fortawesome/free-solid-svg-icons/faPlus"
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
+import {
+    ZoomToggle,
+    AbsRelToggle,
+    HighlightToggle,
+    FilterSmallCountriesToggle,
+} from "grapher/controls/Controls"
+import { ScaleSelector } from "grapher/controls/ScaleSelector"
 
 @observer
 export class ChartTab extends React.Component<{
@@ -20,6 +31,107 @@ export class ChartTab extends React.Component<{
     grapherView: GrapherView
     bounds: Bounds
 }> {
+    @computed get controlsRowControls() {
+        const controls: JSX.Element[] = []
+
+        const { grapher } = this.props
+        const onDataSelect = action(() => (grapher.isSelectingData = true))
+
+        if (grapher.tab === "chart") {
+            const yAxis =
+                (grapher.isStackedArea && grapher.stackedAreaTransform.yAxis) ||
+                (grapher.isStackedBar && grapher.stackedBarTransform.yAxis) ||
+                (grapher.isLineChart && grapher.lineChartTransform.yAxis) ||
+                ((grapher.isScatter || grapher.isTimeScatter) &&
+                    grapher.scatterTransform.yAxis)
+
+            yAxis &&
+                yAxis.scaleTypeOptions.length > 1 &&
+                controls.push(
+                    <ScaleSelector
+                        key="scaleSelector"
+                        scaleTypeConfig={yAxis}
+                        inline={true}
+                    />
+                )
+
+            grapher.canAddData &&
+                !grapher.hasFloatingAddButton &&
+                !grapher.hideEntityControls &&
+                controls.push(
+                    <button
+                        type="button"
+                        onClick={() => onDataSelect()}
+                        key="grapher-select-entities"
+                        data-track-note="grapher-select-entities"
+                    >
+                        {grapher.isScatter || grapher.isSlopeChart ? (
+                            <span className="SelectEntitiesButton">
+                                <FontAwesomeIcon icon={faPencilAlt} />
+                                {`Select ${grapher.entityTypePlural}`}
+                            </span>
+                        ) : (
+                            <span>
+                                <FontAwesomeIcon icon={faPlus} />{" "}
+                                {grapher.addButtonLabel}
+                            </span>
+                        )}
+                    </button>
+                )
+
+            grapher.canChangeEntity &&
+                !grapher.hideEntityControls &&
+                controls.push(
+                    <button
+                        type="button"
+                        onClick={() => onDataSelect()}
+                        key="grapher-change-entities"
+                        data-track-note="grapher-change-entity"
+                        className="ChangeEntityButton"
+                    >
+                        <FontAwesomeIcon icon={faExchangeAlt} /> Change{" "}
+                        {grapher.entityType}
+                    </button>
+                )
+
+            grapher.isScatter &&
+                grapher.hasSelection &&
+                controls.push(<ZoomToggle key="ZoomToggle" grapher={grapher} />)
+
+            const absRelToggle =
+                (grapher.isStackedArea && grapher.canToggleRelativeMode) ||
+                (grapher.isScatter &&
+                    grapher.scatterTransform.canToggleRelativeMode) ||
+                (grapher.isLineChart &&
+                    grapher.lineChartTransform.canToggleRelativeMode)
+            absRelToggle &&
+                controls.push(
+                    <AbsRelToggle key="AbsRelToggle" grapher={grapher} />
+                )
+
+            grapher.isScatter &&
+                grapher.highlightToggle &&
+                controls.push(
+                    <HighlightToggle
+                        key="highlight-toggle"
+                        grapher={grapher}
+                        highlightToggle={grapher.highlightToggle}
+                    />
+                )
+        }
+
+        grapher.isScatter &&
+            grapher.hasCountriesSmallerThanFilterOption &&
+            controls.push(
+                <FilterSmallCountriesToggle
+                    key="FilterSmallCountriesToggle"
+                    grapher={grapher}
+                />
+            )
+
+        return controls
+    }
+
     @computed get layout() {
         const that = this
         return new ChartLayout({
@@ -31,6 +143,9 @@ export class ChartTab extends React.Component<{
             },
             get bounds() {
                 return that.props.bounds
+            },
+            get renderControlsRow() {
+                return that.controlsRowControls.length > 0
             },
         })
     }
@@ -98,7 +213,10 @@ export class ChartTab extends React.Component<{
 
     render() {
         return (
-            <ChartLayoutView layout={this.layout}>
+            <ChartLayoutView
+                layout={this.layout}
+                controlsRowControls={this.controlsRowControls}
+            >
                 {this.renderChart()}
             </ChartLayoutView>
         )
